+++ HISTORY.md
@@ -14,6 +14,7 @@ X.Y.Z | (goals)
   * Added magic method field `clone()` to allow for creating a duplicate copy of `Vec`
   * Added first `example`: `points.c`
   * Added `vec_free()` to free `vec` elements
+  * Added `vec_destroy()` as a shortcut to `vec_free()` then `vec_delete()`
 
 
 0.1.0 | 07-11-2020